pipeline/requirements.txt
# Ontario Health Data Pipeline - Python Dependencies

# Snowflake connector
snowflake-connector-python[pandas]>=4.0.0

# HTTP requests
requests>=2.28.0

# Data manipulation
pandas>=2.0.0

# Web scraping
beautifulsoup4>=4.12.0

# dbt for Snowflake
dbt-snowflake>=1.11.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0

